add notification and chaned gemfile

+++ config/deploy.rb
@@ -8,7 +8,7 @@ set :use_sudo, false
 # for unicorn
 set :unicorn_binary, "/opt/ruby-enterprise-1.8.7-2010.02/bin/unicorn"
 set :unicorn_config, "#{current_path}/config/unicorn.rb"
-set :unicorn_pid, "#{current_path}/tmp/pids/unicorn.pid"
+set :unicorn_pid, "#{shared_path}/pids/unicorn.pid"
 
 role :web, "itjob.fm"                          # Your HTTP server, Apache/etc
 role :app, "itjob.fm"                          # This may be the same as your `Web` server
@@ -22,26 +22,22 @@ namespace :deploy do
   task :init_project do
     run "cd #{release_path}; cp #{shared_path}/config/database.yml #{release_path}/config/database.yml"
     run "cd #{release_path}; ln -s #{shared_path}/uploads #{release_path}/public"
-
-    run "cd #{release_path}; bundle install"
-    run "cd #{release_path}; rake db:create RAILS_ENV=production"
-    run "cd #{release_path}; rake db:migrate RAILS_ENV=production"
   end
 
   task :start, :roles => :app, :except => { :no_release => true } do
-    run "cd #{current_path} && #{try_sudo} #{unicorn_binary} -c #{unicorn_config} -E #{rails_env} -D"
+    run "cd #{current_path} && #{try_sudo} #{unicorn_binary} -c #{unicorn_config} -E production -D"
   end
 
   task :stop, :roles => :app, :except => { :no_release => true } do
-    run "#{try_sudo} kill `cat #{unicorn_pid}`"
+    run "#{try_sudo} kill `cat #{unicorn_pid}`" if File.exist?(unicorn_pid)
   end
 
   task :graceful_stop, :roles => :app, :except => { :no_release => true } do
-    run "#{try_sudo} kill -s QUIT `cat #{unicorn_pid}`"
+    run "#{try_sudo} kill -s QUIT `cat #{unicorn_pid}`" if File.exist?(unicorn_pid)
   end
 
   task :reload, :roles => :app, :except => { :no_release => true } do
-    run "#{try_sudo} kill -s USR2 `cat #{unicorn_pid}`"
+    run "#{try_sudo} kill -s USR2 `cat #{unicorn_pid}`" if File.exist?(unicorn_pid)
   end
 
   task :restart, :roles => :app, :except => { :no_release => true } do
